Title: 낮은 지점들, Time: 78ms, Memory: 1MB, Status: Wrong Answer - CodeTree

+++ 240116/낮은 지점들/lowest-points.cpp
@@ -3,7 +3,7 @@
 using namespace std;
 
 int main() {
-    unordered_map<int, int> hashmap;
+    unordered_map<long long int, long long int> hashmap;
     int n, x, y, sum = 0;
     cin >> n;
     for(int i = 0; i < n; i++)
